+++ spec/monad_axioms.rb
@@ -34,24 +34,5 @@ shared_examples 'a Monad' do
       id1.should == id2
     end
   end
-
-  describe '#map functor' do
-    add100 = lambda { |value| value + 100 }
-    it 'on value types, returns the transformed value, wrapped in the Monad' do
-      res = monad.unit(1).map {|v| add100.call(v) }
-      res.should == monad.unit(101)
-
-      res = monad.unit(1).map {|v| monad.unit(v + 100) }
-      res.should == monad.unit(101)
-    end
-
-    it 'on enumerables, returns the transformed collection, wrapped in the Monad' do
-      res = monad.unit([1,2,3]).map {|v| add100.call(v) }
-      res.should == monad.unit([101,102,103])
-
-      # The following does not work... not sure whether it should
-      # res = monad.unit([1,2,3]).map {|v| monad.unit(v + 100) }
-      # res.should == monad.unit([101,102,103])
-    end
-  end
 end
+